Add proper support for Foreign Keys Removing.
For `DROP FOREIGN KEY constraint_name` with pt-online-schema-change
requires specifying `_constraint_name` rather than the real
constraint_name. This is due to a limitation in MySQL:
pt-online-schema-change adds a leading underscore to foreign key
constraint names when creating the new table.

Additionally this add specs to ensure that the addition and deletion
of foreign keys works correctly.

More details about this:
https://www.percona.com/blog/2017/03/21/dropping-foreign-key-constraint-using-pt-online-schema-change-2/